Update notificacion
Actualizacion de notificacion de estados y categoria

+++ Hotel_CLV/web/pages_admin/categorias_lista.jsp
@@ -11,6 +11,12 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@include file="template/header_admin.jsp" %>
 <!DOCTYPE html>
+<script>
+    function confirmarEliminacion() {
+        var confirmacion = confirm("¿Estás seguro de que deseas eliminar este registro?");
+        return confirmacion; // Devolver true si el usuario hace clic en OK, de lo contrario, false
+    }
+</script>
 <div id="page-wrapper">
     <div class="container-fluid">
         <div class="row">
@@ -49,7 +55,7 @@
                                         <a href="categorias_editar.jsp?editar=true&id=<%= categoria.getCatId()%>" class="btn btn-warning"> <i class="fa fa-edit" title="Editar" name="editar"></i></a>
                                         <!--eliminar update=":tabMostrar"-->
                                         <% if ("ADMINISTRATIVO".equals((String) session.getAttribute("perfil"))) {%>
-                                        <a href="Mantenimiento_admin/crudcategorias_eliminar.jsp?eliminar=true&id_cate=<%= categoria.getCatId()%>"class="btn btn-danger"><i class="fa fa-trash" title="Eliminar" name="eliminar"></i></a>
+                                        <a href="Mantenimiento_admin/crudcategorias_eliminar.jsp?eliminar=true&id_cate=<%= categoria.getCatId()%>"class="btn btn-danger" onclick="return confirmarEliminacion();"><i class="fa fa-trash" title="Eliminar" name="eliminar"></i></a>
                                             <% }%>
                                     </td>
                                 </tr>
